requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.0.0
langdetect==1.0.9
python-dotenv==1.0.0